+++ test/api/test_rotate_thread.cpp
@@ -97,7 +97,7 @@ TEST_P(RotateThread, VectorPlusScalarWithPadding) {
   vector_plus_scalar_test(1);
 }
 
-INSTANTIATE_TEST_SUITE_P(, RotateThread, testing::Values(1, 2, 4, 8),
+INSTANTIATE_TEST_SUITE_P(, RotateThread, testing::Values(1, 2, 3, 4, 6, 8),
                          testing::PrintToStringParamName());
 
 TEST(RotateThreadNotImplemented, InPlace) {